Rename GenericWorkHelper to Sufia::CollectionsHelper
